+++ src/components/Text/Text.tsx
@@ -5,8 +5,9 @@ import { TimeZone } from '@grafana/schema';
 import { Alert, useStyles2 } from '@grafana/ui';
 import { TestIds } from '../../constants';
 import { generateHtml } from '../../helpers';
-import { Styles } from '../../styles';
-import { PanelOptions } from '../../types';
+import { Styles } from './Text.styles';
+import { PanelOptions, RowItem } from '../../types';
+import { Row } from '../Row';
 
 /**
  * Properties
@@ -60,9 +61,9 @@ export interface Props {
  */
 export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, replaceVariables, eventBus }) => {
   /**
-   * Generated html
+   * Generated rows
    */
-  const [html, setHtml] = useState<string[]>([]);
+  const [rows, setRows] = useState<RowItem[]>([]);
 
   /**
    * Generate html error
@@ -85,9 +86,22 @@ export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, rep
    * HTML
    */
   const getHtml = useCallback(
-    (data: any, content: string) =>
-      generateHtml({ data, content, helpers: options.helpers, timeRange, timeZone, replaceVariables, eventBus }),
-    [eventBus, options.helpers, replaceVariables, timeRange, timeZone]
+    (data: any, content: string) => {
+      return {
+        ...generateHtml({
+          data,
+          content,
+          helpers: options.helpers,
+          timeRange,
+          timeZone,
+          replaceVariables,
+          eventBus,
+          options,
+        }),
+        data,
+      };
+    },
+    [eventBus, replaceVariables, timeRange, timeZone, options]
   );
 
   useEffect(() => {
@@ -105,7 +119,12 @@ export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, rep
          */
         const { html, unsubscribe } = getHtml({}, options.defaultContent);
 
-        setHtml([html]);
+        setRows([
+          {
+            html,
+            data: {},
+          },
+        ]);
         unsubscribeFn = unsubscribe;
       } else {
         /**
@@ -132,7 +151,12 @@ export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, rep
            * For every row in data frame
            */
           const rows = data.map((row) => getHtml(row, options.content));
-          setHtml(rows.map(({ html }) => html));
+          setRows(
+            rows.map(({ html, data }) => ({
+              html,
+              data,
+            }))
+          );
 
           /**
            * Call unsubscribe for all rows
@@ -149,7 +173,7 @@ export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, rep
            * For whole data frame
            */
           const { html, unsubscribe } = getHtml({ data }, options.content);
-          setHtml([html]);
+          setRows([{ html, data }]);
 
           unsubscribeFn = unsubscribe;
         }
@@ -187,12 +211,14 @@ export const Text: React.FC<Props> = ({ options, frame, timeRange, timeZone, rep
 
   return (
     <>
-      {html.map((html, index) => (
-        <div
+      {rows.map((row, index) => (
+        <Row
           key={index}
+          item={row}
           className={className}
-          dangerouslySetInnerHTML={{ __html: html }}
-          data-testid={TestIds.text.content}
+          afterRender={options.afterRender}
+          eventBus={eventBus}
+          replaceVariables={replaceVariables}
         />
       ))}
     </>