+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add ESLint deprecation check (#203)
 - Add custom code option which is called after content is ready (#231)
 - Add option to enabled/disable wrapping to support empty lines in HTML (#235)
+- Add handlebars (startsWith, endsWith, match) (#211)
 
 ## 4.1.0 (2023-07-16)